+++ login-workflow/src/screens/EulaScreen/EulaScreen.tsx
@@ -202,9 +202,9 @@ export const EulaScreen: React.FC<EulaScreenProps> = (props) => {
             initialCheckboxValue={eulaAccepted}
             onEulaAcceptedChange={updateEulaAcceptedStatus}
             WorkflowCardActionsProps={workflowCardActionsProps}
-            errorDisplayConfig={errorDisplayConfig}
             refreshConfig={refreshConfig}
             {...otherEulaScreenProps}
+            errorDisplayConfig={errorDisplayConfig}
         />
     );
 };